feat: #107 상단 네비게이션 style 파일 분리
